requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
pytesseract==0.3.8
Pillow==8.3.1
scikit-image==0.18.3
matplotlib==3.4.3
requests==2.26.0
exifread==2.3.2
